requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-location-field==2.1.0
django-mathfilters==1.0.0
django-phonenumber-field==4.0.0
django-twilio==0.13.0.post1
gunicorn==20.0.4
idna==2.9
phonenumbers==8.12.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
twilio==6.37.0
urllib3==1.25.8
